Starting to tidy up. -Split all sections out to separate files - Did more mobile tidying
