+++ package.json
@@ -31,6 +31,7 @@
     "react-scripts": "3.2.0",
     "react-starfield-animation": "^1.0.1",
     "react-timezone": "^2.4.0",
+    "stats.js": "^0.17.0",
     "styled-components": "^5.0.0-beta.9",
     "three": "^0.139.2",
     "webfontloader": "^1.6.28"
