+++ src/components/page-sections/index.js
@@ -1,0 +1,7 @@
+export { ApplySection } from "./ApplySection";
+export { HomeSection } from './HomeSection';
+export { ChatSection } from "./ChatSection";
+export { MetaverseSection } from "./MetaverseSection";
+export { ScheduleSection } from "./ScheduleSection";
+export { SpeakersSection } from "./SpeakerSection";
+export { WorkshopsSection } from "./WorkshopSection";
