+++ src/index.js
@@ -2,87 +2,93 @@ import React from "react";
 import "./index.css";
 import App from "./App";
 import * as serviceWorker from "./serviceWorker";
-import { ChakraProvider, extendTheme } from "@chakra-ui/react";
-import * as ReactDOMClient from 'react-dom/client';
+import { ChakraProvider, CSSReset, extendTheme } from "@chakra-ui/react";
+import * as ReactDOMClient from "react-dom/client";
 
-const container = document.getElementById('root');
+const container = document.getElementById("root");
 // Create a root
 const root = ReactDOMClient.createRoot(container);
-
 
 const theme = extendTheme({
   styles: {
     global: {
-      'html, body': {
+      "html, body": {
         fontFamily: '"Exo 2", sans-serif',
-        fontSize: '16px',
+        fontSize: "16px",
         fontWeight: 300,
-        bg: 'linear-gradient(0deg, rgba(41,2,80,1) 0%, rgba(25,0,50,1) 40%)',
+        bg: "linear-gradient(0deg, rgba(41,2,80,1) 0%, rgba(25,0,50,1) 40%)",
       },
+      // "body, body *": {
+      //   outline: "1px solid red",
+      // },
       a: {
-        color: '#FF61E6',
-        textDecoration: 'none',
-        transition: 'color 0.2s ease',
+        color: "#FF61E6",
+        textDecoration: "none",
+        transition: "color 0.2s ease",
         _hover: {
-          color: '#76EBF2',
-          textDecoration: 'none'
+          color: "#76EBF2",
+          textDecoration: "none",
         },
-        '&.chakra-link': {
-          color: '#FF61E6',
+        "&.chakra-link": {
+          color: "#FF61E6",
           _hover: {
-            color: '#76EBF2',
-            textDecoration: 'none'
+            color: "#76EBF2",
+            textDecoration: "none",
           },
-        }
+        },
       },
-      'h1': {
-        color: '#fff',
-        fontSize: '4vmax',
+      h1: {
+        color: "#fff",
+        fontSize: "4vmax",
         fontWeight: 700,
-        '& span': {
-          color: '#fff',
-          fontSize: '4vmax',
+        "& span": {
+          color: "#fff",
+          fontSize: "4vmax",
           fontWeight: 700,
-        }
+        },
       },
       h2: {
-        color: '#fff',
-        fontSize: '4vmax',
+        color: "#fff",
+        fontSize: "4vmax",
         fontweight: 500,
-        textShadow: '0 0 10px rgba(0, 0, 0, 0.6)',
+        textShadow: "0 0 10px rgba(0, 0, 0, 0.6)",
       },
       h3: {
-        fontSize: '1.5vmax',
+        fontSize: "1.5vmax",
         fontWeight: 500,
+        "& + p": {
+          fontSize: "1vmax",
+          fontWeight: 500,
+          mb: 1,
+        },
       },
       h4: {
-        fontSize: '1vmax',
+        fontSize: "1vmax",
         fontWeight: 500,
       },
       p: {
-        fontSize: {base: '2.2vmin', md: '.9vmax'},
-        textShadow: '0 0 5px rgba(0, 0, 0, 0.6)',
+        fontSize: { base: "2.6vmin", md: ".9vmax" },
+        textShadow: "0 0 5px rgba(0, 0, 0, 0.6)",
       },
-      '.__content__body': {
-        'p:first-of-type': {
-          fontSize: '1vmax',
-          fontWeight: 500
+      ".__content__body": {
+        "p:first-of-type": {
+          fontSize: {base: '2.6vmin', md: "1vmax"},
+          fontWeight: 500,
         },
-        '&--no-firstof': {
-          'p': {
-            fontSize: '0.9vmax',
+        "&--no-firstof": {
+          p: {
+            fontSize: "0.9vmax",
             fontWeight: 500,
-            '& + p': {
-              fontWeight: 300
-            }
-          }
-        }
+            "& + p": {
+              fontWeight: 300,
+            },
+          },
+        },
       },
       // Gradients
       ".gradient": {
         display: "inline-block",
-        background:
-          "linear-gradient(90deg, #FF61E6 -29.22%, #7C56FF 107.53%)",
+        background: "linear-gradient(90deg, #FF61E6 -29.22%, #7C56FF 107.53%)",
         backgroundClip: "text",
         // WebkitTextFillColor: "transparent",
         // textFillColor: "transparent",
@@ -112,14 +118,28 @@ const theme = extendTheme({
         filter: "drop-shadow(0 0 5px rgba(0,0,0,0.6))",
       },
       ".highlight": {},
+      ".fest-dates": {
+        d: "inline-flex",
+        width: "100%",
+        color: "#FF61E6",
+        fontSize: { base: "2.2vmin", md: "0.7vmax" },
+        fontWeight: 700,
+        justifyContent: "right",
+        pr: 0.5,
+        transform: { base: "translateY(7px)", lg: "translateY(9px)", '2xl': "translateY(15px)" },
+        zIndex: 2001,
+      },
     },
   },
 });
 
 root.render(
+  <>
+    <CSSReset />
   <ChakraProvider theme={theme}>
     <App />
-  </ChakraProvider>
+    </ChakraProvider>
+    </>
 );
 
 // If you want your app to work offline and load faster, you can change
